Restructure module.json due to misunderstanding
Originally I misread Louis' 2015-11-06 email and thought he'd said we
need to have the links to the quizzes on the video page under the
resources as opposed to on the phase page, but actually he just meant
that he hasn't changed the phase template in his patch and that we'd
have to do that ourselves.

I have repositioned the location of the quiz definition in module.json
accordingly.