Display question options on quiz page
The questions of a quiz now show each multiple choice option. To do this
some modification of the JSON format was required. The structure holding
the options was renamed from "answers" to "options" because it conveys
the meaning better, and each option now contains two attributes, letter
and text, rather than the key being the letters and the value the text,
which was difficult to iterate through.