Add Helpers/BadgeUtils.rb
This file will hold helper functions for dealing with badges - currently, it loads them all and gives them a numerical ID similarly to how the phases are loaded.
This helper has been included in main.rb, allowing it's functions to be called as if they were declared in that file.