Differentiate between String ID in JSON and calculated position id
Previously all resources were identified in order of appearance,
but now we also have a string identifier defined in the json.

This changes behaviour to not overwrite the string ID, and to differentiate
between that id and the positional id, such that flippd behaves as it did before (i.e.
next, prev links can be found by getting the prev and next page by order of apperarance)